Get working example of column chooser with descriptions
