+++ src/app/components/list-view-grid/list-view-grid-demo.component.html
@@ -3,6 +3,7 @@
     <sky-list-toolbar-secondary-actions>
       <sky-list-toolbar-secondary-action>
         <button type="button" (click)="openColumnSelector()">
+          Choose columns
         </button>
       </sky-list-toolbar-secondary-action>
     </sky-list-toolbar-secondary-actions>
