+++ src/modules/column-selector/column-selector-context.ts
@@ -1,0 +1,10 @@
+export class SkyColumnSelectorContext {
+  public columns: Array<SkyColumnSelectorModel>;
+  public selectedColumnIds: Array<string>;
+}
+
+export class SkyColumnSelectorModel {
+  public id: string;
+  public heading: string;
+  public description: string;
+}
