+++ src/modules/column-selector/column-selector-modal.component.html
@@ -11,7 +11,7 @@
         [placeholder]="'grid_column_picker_search_placeholder' | skyResources">
       </sky-list-toolbar>
 
-      <sky-list-view-checklist label="heading">
+      <sky-list-view-checklist label="heading" description="description">
       </sky-list-view-checklist>
     </sky-list>
   </sky-modal-content>
