--- src/modules/grid/column-selector-context.ts
@@ -1,6 +1,0 @@
-import { SkyGridColumnModel } from './grid-column.model';
-
-export class SkyColumnSelectorContext {
-  public columns: Array<SkyGridColumnModel>;
-  public selectedColumnIds: Array<string>;
-}
