+++ src/modules/grid/index.ts
@@ -3,5 +3,3 @@ export * from './grid.component';
 export * from './grid-column.component';
 export * from './grid-cell.component';
 export * from './grid-column.model';
-export * from './column-selector-modal.component';
-export * from './column-selector-context';
